requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Migrate
Flask-Limiter
psycopg2-binary
python-dotenv
Werkzeug